Refactor components for post cards
