+++ src/components/department-head/RequestsInProgressComponent.jsx
@@ -1,26 +1,17 @@
 import * as React from 'react';
-import { StyleSheet, View } from 'react-native';
-import { Text } from 'react-native-paper';
+
 import CurrentPostRequestCard from './CurrentPostRequestCard';
+import BaseRequestComponent from '../BaseRequestComponent';
+
 
 const RequestsInProgressComponent = ({ posts }) => {
     return (
-        <View>
-            <Text style={styles.textStyle} variant="titleLarge">Requests in Progress ({posts.length})</Text>
-            {posts.map((post, index) => (
-                <CurrentPostRequestCard key={index} {...post} />
-            ))}
-        </View>
+        <BaseRequestComponent 
+            posts={posts}
+            CardComponent={CurrentPostRequestCard}
+            title="Requests in progress"
+        />
     )
 };
-
-const styles = StyleSheet.create({
-    textStyle: {
-        fontWeight: "bold",
-        fontSize: 30,
-        margin: 20,
-    },
-
-})
 
 export default RequestsInProgressComponent;
